+++ src/components/Badge/Badge.stories.tsx
@@ -1,0 +1,19 @@
+/*
+ * Copyright (c) 2021 AccelByte Inc. All Rights Reserved.
+ * This is licensed software from AccelByte Inc, for limitations
+ * and restrictions contact your company contract manager.
+ */
+
+import React from "react";
+import { Meta, Story } from "@storybook/react/types-6-0";
+import { Badge, BadgeProps } from ".";
+
+export default {
+  title: "Components/Badge",
+  component: Badge,
+} as Meta;
+
+const Template: Story<BadgeProps> = (args) => <Badge {...args} />;
+
+export const Default = Template.bind({});
+Default.args = { text: "Badge" };