+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.30]
+### Added
+- `success` appearance and `dataQa` prop in `Button`
+- `Badge` story
+
+### Removed
+- `qaProps` prop from `StyledRadio`
+
+### Fixed
+- Change `data-qa` to `data-qa-id` in `StyledRadio`
+
 ## [0.1.29]
 ### Fixed
 - Only render `InlinePopover` in `ValidFieldText` if the value for `popoverContent` prop is set